Debug State
In the future, N_SPEED_STEPS should be raised back to 3000 from the
current 100.
To do this, the planners will only send the changes to their
LaneGraphTrajectory, instead of the whole trajectory.
This adds loads of debug output to stdout to make this change easier,
while keeping the example in a working state.