Improved reading of messages and debug output
When reading LaneGraphTrajectoryChange messages from other controllers,
it will compare the timestamp of the message to the current time, and
write it into a more fitting index in the previous_vehicle_buffer.

Debug output right now is excessive for general use and should be removed again.
However, it can help to see, if the trajectories of other vehicles are
received correctly.