Changed to sending LaneGraphPoint's
This pretty much GUARANTEES that late-joiners still get all information
that they need.

Since LaneGraphPoints use the vehicle_id and the estimated_arrival time
as keys, all messages are unique (except for changes to existing points,
but that's fine).
Because we only look at 3000 points, we set the History QoS of the
LaneGraphPoint writers to KeepLast(3000), so that the last 3000 messages
are kept.